Moved renderer related files to /services/renderer
