+++ packages/app/src/services/renderer/PreProcessor/XssFilter.js
