+++ packages/app/src/services/renderer/interceptor/detach-code-blocks.js
