+++ packages/app/src/services/renderer/markdown-it/emoji-mart-data.ts
