+++ packages/app/src/services/renderer/markdown-it/mathjax.js
